requirements.txt
loguru==0.7.3
numpy==2.3.1
onnxruntime==1.22.0
onnxsim==0.4.36
Pillow==11.3.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
